[refactor] filter 패키지로 정리 #9
